refactor: made NumberHandleUtils with two methods which were in Application before (convertIntToArray, isValidNum)
